Show correct root account locked status in reconfig mode (#1507940)
It is currently possible (and expected on ARM images) to
re-enable a locked root account by setting a new root password
if the root password spoke is running in Initial Setup reconfig mode.

Unfortunately the spoke status in this case just says the password
is set without mentioning that the root account is locked.

So fix the root password spoke status to mention that the root account is locked
but can be re-enabled by setting a new root password.

Resolves: rhbz#1507940