Only check space during a tui kickstart if ksprompt is enabled
The problem is that if it isn't enabled it hits the check, which guesses
at the size (3000 MB) since depsolving isn't done yet and it crashes
(which is a separate problem).

When doing unattended ks installs it doesn't make sense to stop at this
point, the size is re-evaluated later, when the depsolve is finished and
it has the correct value to check against.