Do not manually create LUKSDevice when unlocking a LUKS format
`populate` will create the device and add it to devicetree
automatically.
This also solves issues with LUKS 2 devices with integrity where
the LUKS device is not direct child of the partition that is
being unlocked. (So it is not possible to manually create it from
Anaconda and manually set its parents.)