Handle missing package errors reported by the install_specs() function
Unlike in the past where we asked users about missing packages and
groups one by one, the new install_specs() function provided by DNF
gather all missing packages and groups/modules and returns them
as a list of strings.

So handle this case and correctly report all the missing items to
the user, which can then decide how to proceed.