buildinstall: fix losetup result (loop device name) parsing
Resolves: rhbz#1589428

If losetup -f finds out that the loop device name has been stolen,
it tries to find another one but we were not able to parse
the device name from its output at this point.

Also add a check on losetup return code and propagate eventual fail to
buildinstall insead of failing to create the image silently.