passing a non existing disk to ignoredisk should not halt the installation
I propose this small change to avoid an installation from exploding simply because a disk to ignore doesn't exist. 
Since it allows globs users can specify ignoredisk --drives=/dev/disk/by-path/*usb* if I wish to never install on a usb disk or something similar. But it shouldn't fail since I'm only excluding and if it doesn't exist, it should be ok.