Add system purpose support
The system purpose functionality is used to give hints to other tools about the
intended usage of the system.

Anaconda plays adds support for setting system purpose values via
kickstart & GUI and forwards them to the syspurpose tool which is
called inside the installed system chroot after the package installation
and configuration phase has been finished.

Also as with other kickstart/spoke combos in Anaconda, the values
in kickstart should reflect in the GUI if possible and kickstart
can thus be used to pre-select system purpose values in the GUI
if needed.

If a value from kickstart does not match to any of the valid values,
it will still be forwarded to the syspurpose tool and shown as
Other (<value>) in the GUI.

The system purpose support in Anaconda is considered fully optional,
so an installation can be completed without setting any system purpose
values in kickstart or GUI. In the same way Anaconda can tolerate
missing list of valid system purpose values and even no syspurpose utility
being available on the installed system.

Resolves: rhbz#1612060