Add support for DBus structures
DBus structure is a dictionary of type Dict[Str, Variant] that maps
field names to variants of field values. However, it might be easier
to work with objects instead of dictionaries.

A data class that describes a DBus structure should be decorated
with the @dbus_structure decorator. Then instances of this class can
be transformed to a DBus structure with the get_structure function, or
a DBus structure can be transformed to an instance of this class with
the apply_structure function.