Make sure dracut can make pmem devices from NFIT-based ramdisks.
UEFI 2.7 and ACPI 6.2 support a feature where the firmware sets up a
ramdisk, which it then boots from.  Once the OS is booted, the memory
the ramdisk is in shows up as "reserved", and the firmware marks it in
the ACPI NFIT table as one of four various kinds of ramdisk.

When linux sees this data, if we have the nfit and nd_* nvdimm drivers
available, it creates /dev/pmemM and /dev/pmemMpN devices.

This patch enables anaconda to use that device as its location for e.g.
netinst.iso .

Signed-off-by: Peter Jones <pjones@redhat.com>