Make pyanaconda.dbus.typing work with Python 3.7 (#1598574)
As reported in RHBZ#1598574, the internals of typing changed in
Python 3.7 such that it's no longer so simple to find the 'base'
type of a type hint (it's not just its `__origin__` any more).
There doesn't appear to be any particularly great fix for this,
but this suggestion from Miro Hrončok seems as good as any other
option we have for now. This should work with both 3.6 and 3.7.

Signed-off-by: Adam Williamson <awilliam@redhat.com>