Use Platform Python where appropriate
Anaconda should explicitly use Platform Python when running
in the installation environment, which is expected to be Platform Python
only. Even though a shebang rewriter should be in place in the
buildsystem, this still needs to be done explicitly to make updates
images work and due to calling the Python binary directly when starting
DBUS modules.

Maintenance and build scripts are still expected to run in
an environment where a "normal" Python will be available so
no changes are needed.