Load base repository location from treeinfo
The main repo can be a subdirectory in the compose. This will enable to
have cdrom installation with multiple repositories with a sane
structure.

Information where the repository is placed is read from the .treeinfo
file. When this file exists then we will look for the known base
repository names and if found the repository url is changed.

Related: rhbz#1610806