Disable treeinfo based repos only once
It is bad user experience to disable treeinfo repositories for every
change so when user enable treeinfo repo and change the base repo again
then it is again disabled. The solution is to disable the repo only for
the first change of the base repository.

This works every time when user enters Source spoke.

Related: rhbz#1610806