Set platform id for DNF
To correctly install modules DNF needs to know the correct platform id.

On an installed system DNF will get it from /etc/os-release but during
an installation run the target system chroot is empty and there is
no /etc/os-release for DNF to parse.

So parse the /etc/os-release from the installation environment (if any)
and set the DNF platform id based on the PLATFORM_ID key.

Related: rhbz#1613296