Warn when repo names are not unique
When there are multiple repositories with the same name then the first
repo is used. This is opposite from our usual kickstart behavior where
the last command is used.

Unfortunately, we can't change this because users might depend on that
so only log this in the package log.

Related: rhbz#1595415