Don't connect to signals of the Network Manager DBus objects (#1582233)
Every proxy of DBus object provided by the Network Manager connects to
the signals and loads all properties. However, we don't use the signals
and we don't access the cached properties, so it is unnecessary.

It is also a workaround for the bug #1582233, because Anaconda can
freeze when these proxies tries to disconnect from the signals.

Resolves: rhbz#1582233