Flags shouldn't process the kernel options
There is no reason why we should process the kernel options in the
flags, because these flags are later set with the Anaconda options
anyway. The kernel options are handled there.

Also, the flag noefi can be removed, because we don't use it and the
flag leavebootorder should be set from the Anaconda option.