Update kernel command line parameters in BLS files
On kernel installation, the kernel-install scripts create a BootLoaderSpec
file and set the kernel command line params to the currently used cmdline.

But this shouldn't be done during system install, when the first kernel is
installed since in that case the BLS will have the cmdline used on install.

In that case Anaconda should update the kernel command line parameters with
the correct values needed to boot the installed system.

Resolves: rhbz#1635180

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>