Don't check for firmware compatibility to enable BootLoaderSpec support
When Power systems are installed on bare-metal and using the OPAL firmware
interface, the Petitboot bootloader is used to parse the bootloader config
and populate the boot menu entries.

Support for the BootLoaderSpec config files was added in Petitboot 1.8.0,
so this is the minimal version needed to rely on Petitboot to parse them.

But there are machines that have an older firmware version. So instead of
not allowing the installation to finish, enable BLS support and have a way
to generate a grub.cfg file that's supported by any Petitboot version.

Resolves: #1635547

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>