Replace helpFile property with help_id
Previously we used to have one help content file per screen,
which no longer works for the new help content, which is using
a single docbook file with anchors.

So replace the helpFile property with a help_id property, which
holds a unique string id corresponding to one screen needing help.

The help id can then be mapped into an anchor in the main docbook file,
or even something else if needed.

Related: rhbz#1593723