Create a class for the Anaconda bus connection
Created a new class for the Anaconda bus connection, that can be
specified by an environmental variable or by a file. This class
is a parent of the default connection.

It is possible to use our DBus support to access the Anaconda DBus
modules without setting up the the environment now.