Let the DBus launcher to set up the modules
The Anaconda DBus launcher is able to set up and clean up everything
related to the DBus modules. It will create a temporary configuration
file, start a DBus session, set up the environment, start the boss
and kickstart modules and wait for the modules to start. It raises
TimeoutError if the modules don't start on time.