Re-generate BLS loader file snippets on live installs
Currently, on live installs, the BLS loader snippets that are
part of the live image itself - in /boot/loader/entries - are
simply rsync'ed into the installed system along with everything
else. This is no good in most cases because they point to
`/boot/vmlinuz-(foo)` and `/boot/initramfs-(foo)` (since the
live image itself does not have a separate /boot partition),
whereas most installs will have a separate /boot partition so
the path specified in the config snippet should not include
`/boot`. (The files are also named for the machine ID that was
used during live image generation, which will not be the machine
ID of the installed system; I'm not sure exactly what the
consequences of this are, probably nothing major but it may be
somehow an issue when the kernel is removed, perhaps.)

To fix this, we should exclude these files from the rsync, and
instead do something to have them re-generated after the rsync
has run. There are a few different ways and places we could do
this, but my choice is to run `kernel-install` in live payload
postInstall. This seems like it should be fairly robust. In
testing it does the job, both BIOS and UEFI installs that I
tested installed cleanly and booted successfully. Note that
`kernel-install` must run *after* `/etc/machine-id` exists in
the install root, or it will bail immediately and do nothing.
Note that at present, the 'Generating rescue image' step in
`install()` actually results in the generation of a machine ID,
so this could be done any time after that, but doing it after
the explicit failsafe machine ID generation in `postInstall()`
seems safest.

Signed-off-by: Adam Williamson <awilliam@redhat.com>
Signed-off-by: Brian C. Lane <bcl@redhat.com>

Related: rhbz#1684542