Don't start TUI main loop if no screens are scheduled (#1646568)
We need to make sure the TUI main loop is only started if some screens
are scheduled.

This (no screens scheduled as there is nothing to do) is a valid
usecase for Initial Setup if the user has already been created,
there is no EULA to show and reconfig mode is not active.

Unfortunately instead of the expected behavior (do nothing and exit)
the main loop would still be started even with no screens, effectively
stopping the boot process while not showing anything on the console.

To avoid this scenario we need to check that there are some screens
scheduled before we start the TUI main loop.

Resolves: rhbz#1646568