Disable BLS config if new-kernel-pkg script is installed
The kernel-install scripts installs kernels and adds BLS snippets in the
/boot/loader/entries directory if new-kernel-pkg isn't present or a BLS
setup was explicility configured (i.e: setting GRUB_ENABLE_BLSCFG=true).

But the kernel package is installed before writing any bootloader config,
so the kernel-install script only rely on the existence of new-kernel-pkg
to choose if a BLS or non-BLS kernel installation will be made.

So if new-kernel-pkg is present and the inst.noblscfg boot option wasn't
set, then kernels will be installed without a BLS snippet but bootloader
config will expect them, leading to an unbootable installed system.

Resolves: rhbz#1654036

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>