Make save button insensitive in LUKS dialog with no input (#1489713)
Previously the save button would be sensitive in the LUKS passphrase
dialog even if no passphrase and passphrase confirmation has been
filled by the user. This by itself will be detected later and
the installation blocked until the passphrase is fixed.

Still, this is bad UX, so make the save button insensitive if both
the passphrase and passphrase confirmation fields are empty.

Resolves: rhbz#1489713