Ensure grubby-deprecated is installed if blscfg is disabled
If blscfg is diasbled (with e.g. inst.noblscfg), we must install
grubby-deprecated or else the installed system will not be able
to update the bootloader configuration properly when the kernel
is updated. This converts `packages` from a class attribute to
a property because this `flags.blscfg` check just doesn't seem
to work if done while setting up a class variable; even when
booting with `inst.noblscfg`, we don't get grubby-deprecated
installed. Implemented in a property like this, it works right.

Signed-off-by: Adam Williamson <awilliam@redhat.com>