Drop the inst.noblscfg option
This option was added to allow installing a system configured without BLS
support. But the option is misleading because even when the bootloader is
configured without BLS, the grubby-deprecated package must be installed
in order to install new kernels and add entries in the bootloader config.

The kernel-install scripts already use the existence of the new-kernel-pkg
to decide if a BLS snippet will be added or not when a kernel is installed
and also is used to configure a bootloader with BLS support or not.

So the inst.noblscfg option doesn't really serve a purpose at this point
and can just be dropped. Instead, a user can choose to avoid configuring
the system with BLS by installing the grubby-deprecated package.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>