Drop the inst.noblscfg option (#1658980)
This option was added to allow installing a system configured without BLS
support. But the option is misleading because even when the bootloader is
configured without BLS, the grubby-deprecated package must be installed
in order to install new kernels and add entries in the bootloader config.

However, that is not supported on RHEL. So the inst.noblscfg option
doesn't really serve a purpose at this point and can just be dropped.

(cherry picked from d87eb1e)

Resolves: rhbz#1658980