Add comment to /etc/default/grub explaining how cmdline can be set with BLS
On a BLS configuration the menu entries aren't defined in the grub.cfg but
instead defined in BLS snippets in the /boot/loader/entries directory. So
re-generating the grub.cfg with grub2-mkconfig has no effect on the menu
entries or their kernel command line parameters.

That's why modifying GRUB_CMDLINE_LINUX in /etc/default/grub and executing
grub2-mkconfig can't be used anymore to modify the entries kernel cmdline.

This can be confusing for users that are used to those steps to modify the
cmdline, so let's add some comments in /etc/default/grub explaining that's
not supported anymore and how the cmdline can be modified on a BLS setup.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>