Document live image and boot iso creation
Document live image and boot iso creation from the point of view
of testing changes in Anaconda, it's packaging and possibly also
related components.

Some changes can only be fully verified by building a live image or
a boot iso and then checking if it works correctly.

Therefore it's important to have a proper guide how to easily build such
image that contains all the relevant changes in Anaconda or other
related software.