Only quit GTK mainloop on GUI quit request (#1643111)
If the user selects the quit action from GUI only quit the GTK mainloop
instead of killing the whole process by calling exit(0).

Initial Setup starts the GUI mainloop and then expects it to always to quit,
so that it can run various post-configuration actions, such as rebooting
the machine on RHEL if EULA has not been accepted.

This change should not influence "normal" non-IS Anaconda runs,
as in such case ending the mainloop should be equivalent to killing
the process.

Resolves: rhbz#1643111