Fix source spoke status for expanded tree on HDD
Translation from device spec to device name have to be done by the udev
not blivet. Blivet is using devicetree to handle this but in this case
the device should be hidden, so it is not part of the devicetree.

Related: rhbz#1502323
(cherry picked from commit 3bc4f3caac490618f7b9f13dad77dacc22525010)