Fix anaconda-live package temporary Requires:
At the moment there is a temporary Requires: in place so that
the anaconda-live package is always installed when needed.

Eventually this will be removed once the anaconda-live package
is explicitly required in the kickstart files used for live image
generation.

Initially this was simply added to the anaconda-core package,
but this turns out to be incorrect, as it transitively makes
the anaconda-core package depend on anaconda-gui. This not only
pulls in unnecessary dependencies, but also fails the compose on
armv7 where dependencies needed by the Anaconda GUI are not available.

So move the temporary Requires for anaconda-live from anaconda-core
to anaconda-gui, which should fix the issue.