Relabel X11/xorg.conf.d directory
The keyboard configuration file is copied from the
/etc/X11/xorg.conf.d/ to the installed system. The problem is that the
installation environment has SELinux label set to var_lib_t but it
should have xserver_etc_t. We don't have problem with this during
installation but it starts to be a problem in the installed system.

We are copying this file from the installation environment by
shutils.copy2 but this will copy the file metadata too.

Resolves: rhbz#1660436