Add support for managing multiple users to the users module
Add a new DBUS api based on a simple list of dicts, where each
dict represents a single user.

The order of the dicts corrresponds to the order in which users
have been found in kickstart (if any).

Also remove the previous initial API for dynamic user management
and adjust the unit tests.