Clean up the space checkers (#1520749)
Remove the reset method and the deficit attribute. They are used only
internally.

Don't calculate the deficit if there is no root device. Otherwise,
anaconda might fail with a traceback.

Remove duplicate code from DirInstallSpaceChecker. We only need to
redefine a way how to calculate the available space. Otherwise, the
code is the same as in FileSystemSpaceChecker.

Resolves: rhbz#1520749