Remove support for deprecated bootloaders
The legacy GRUB and Yaboot bootloaders are not used anymore, instead the
GRUB2 bootloader is used to support the platforms that were using these.

This commit only removes the EFIGRUB1 and Yaboot classes, but leaves the
GRUB class since it's the parent class of GRUB2. A follow-up change will
remove the GRUB base class and move the shared code to the GRUB2 class.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>