Restore RHEL-specific stylesheet data (#1638383)
Update the color for RHEL, it should be same as for Fedora. Set
the color theme_bg_color to the RHEL color and use it to define
the style.

Don't repeat the background image of the navigation bar.

Remove the useless style for #nav-box.

(inspired by a commit 6e6ba58)

Resolves: rhbz#1638383