Add the GetResult method
The GetResult method returns a result of a task. It will raise the
NoResultError exception if the task doesn't provide a result or the
result is not publishable on DBus. A task doesn't provide a result if
it has failed.

If you want to provide a result of a task, return the result in the
run method of the task and create a new class that will override the
convert_result method of the TaskInterface class. Publish the task
with this interface class.

For example:

  @dbus_class
  class MyTaskInterface(TaskInterface):

    @staticmethod
    def convert_result(value):
      return get_variant(Int, value)