Recreate the BLS entries when using liveimg
If the source image (eg. tar) had no /boot mountpoint the BLS entry will
not work on an installed system with /boot as a separate partition. So
this patch does a few things to correct this:

 * Always make a new machine-id at the end of install
 * Fix the tar --exclude lines to actually exclude the paths
 * Remove any existing BLS entries from /boot/loader/entries/

The machine-id creation has to be done before calling grub2-set-default,
grub2-mkconfig, or kernel-install so doing it in post_install is too late.

Resolves: rhbz#1684542