efi: don't include the grub2-pc package on EFI installs
The GRUB2() class adds the grub2 package to the list of packages to be
installed and the EFIGRUB() class inherits this which leads to grub2-pc
package being installed even on EFI installs. That's wrong because this
isn't needed for EFI and also causes a broken /etc/grub2.cfg symlink to
be created.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>