Add API for easy admin user detection
Make it easy to detect if at least one admin user exists.

By exists we mean Anaconda plans to create it based on user input so far.

This *does not* mean it will actually be created on the target system
due to the way Anaconda handles failures during user creation
(logging a warning & continuing).