Deprecate DBusObjectObserver
The class DBusObjectObserver is deprecated. Use a DBus proxy to access
a DBus object and the class DBusObserver to monitor a DBus service.
This combination provides the same functionality as DBusObjectObserver,
but in a better and understandable way.

This will also prevent bugs caused by access to a proxy of the object
observer through the proxy property.

We use almost always the get_observer method of DBusServiceIdentifier
to get an object observer, so let's issue the warning from there.