Remove set_up_bootloader from InstallerStorage
Move the code from the set_up_bootloader method to the BlivetExecutor
class and use a new function setup_bootloader instead of the executor.

Don't set up the bootloader after autopart. The bootloader will be
setup again in the _setup_bootloader method, so we can remove this
call. Update the InteractiveAutoPartitioningTask to call the method
_setup_bootloader as well.