Remove askmethod flag
Based on the documentation this option is not used for a long time so
remove this from the code completely.

This code was removed in the past in the commit:

d4f73641f31f8c17caa3f75766ec4de607a39785

And reintroduced in the commit because of bug bz#889887:

ae9e2239712eefcda683bc625343697085d53152

This reintroducing wasn't in the documentation and the justification was
really weak. We can add this back if there will be a reason to but with
a meaningful way, not by adding old deprecated option which was
designed for something else.