payload: Mark DefaultEnvironment API as temporary
This property presents --default attribute of the %packages section.
The attribute doesn't work as it is expected and the problem is crawling
between Anaconda and Pykickstart.

Right now the behavior is implemented in a way that when --default is
set it will enforce first (default) environment and it will generate an
empty %packages section. This behavior is wrong from multiple points.

GUI: In partial kickstart installation user will see the default (first)
environment selected and he can change that and set additional groups
they want to install. The problem is that even when the user set the
environment it will still install only the default environment, in other
words the default environment is forced without presenting this to a
user. Groups marked by the user will be installed but won't be presented
in the output generated kickstart.

Automated installation: You can write groups/modules and packages in the
kickstart file but since there is --default set then these packages and
groups are installed but they are not present in the output generated
kickstart.

Right now this API will be here but later the plan is it will be moved
to the DNF handler module as select default environment method. This
method will only select the default environment but it won't force it
later. The pykickstart code should be changed to generate the packages
section even with this attribute enabled.