Remove no longer needed kickstart command overrides
The user related configuration is now executed in DBUS tasks and
the UI uses the Users DBUS module instead of kickstart,
so we can start dropping the overrides Anaconda has for user
configuration related commands.

We still need to keep at least one user related command
due to the __str__ method used to compile the output kickstart,
but we can drop the rest:
- user
- sshkey
- group
And the execute() method of the rootpw command.