Fix bad assumption about base repo identification (#1691832)
This works for ISO and I have a feeling this is not true for repo files.

Fedora don't have naming of the repository that we are able to identify
base repository so when we don't found base repository on the ISO we
will just use root of the ISO as base repository and see how it goes.

Our tests for valid ISO where too strict about this. Right now we will
look for base repo and if that is not found we will look on repo which
has repository in the root.

Resolves: rhbz#1691832