Don't define the __eq__ method of DBusData
The definition of the __eq__ method makes objects unhashable unless
the __hash__ method is defined as well. However, we cannot implement
this method, because the hash is supposed to be immutable and the
data of DBusData are mutable.

Let's use directly the function compare_data to compare data of two
objects.