Fix non-root dir of install tree HD installation (#1689194)
Installation didn't worked if you placed an install tree on the other
place than / of the partition. This was because `method.dir` was not
correctly propagated in this case.

The reason is, before installation from an install tree was supported
only the ISO installation was supported from HD. In that case you
would use `method.dir` to find the ISO but the path to the installation
source would be always root of the mounted image.
This does not work for install tree because we will mount partition and
then have to use the path to adjust installation source on that mount
point.

Resolves: rhbz#1689194