Fix non-root dir of install tree HD installation (#1689194)
Installation didn't worked if the install tree was placed on the other
place than / of the partition. This was because `method.dir` was not
correctly propagated in this case.

In the past the only supported installation from HD was with ISO image.
In that case you would use `method.dir` to find the ISO but path to
the installation source was always root of the mounted image.
This does not work for the install tree. Installation path have to be
adjusted from the mount point to the specified path.

Resolves: rhbz#1689194
(cherry picked from commit 559ad599790e8791b70a8cf0e95b002fb0035274)