Add root parameter to execInSysroot utility function
Make it possible to set the sysroot for the execInSysroot function
so that callers can specify what sysroot should be specified.

This is quite important for installation tasks as they run
in the context of the process of the given DBUS module where
getSysroot() might not always return the correct sysroot path.