Improve post install tools configuration
Improve how post install tool configuration is done
by adding a new property tracking if post install tools are enabled
and by changing the default enabling Initial Setup.

The PostInstallToolsEnabled property

Add a property that tracks "firstboot --disable" in kickstart
and sets the PostInstallToolsEnabled property of the services
module accordingly. The property can of course also be set
via DBus.

The aim of this addition is to make it clear if post installation
tools (such as Gnome Initial Setup) will be marked as disabled.
We also make sure the property is set to True if "firstboot --disable"
is is present in the installation kickstart or when set via DBus.

Disable Initial Setup by default

Instead of trying to enable Initial Setup by default and then
overriding this for all automated (kickstart based) installations,
let's just always disable Initial Setup by default and always
enable it for interactive (no user provided kickstart) installations.

This way the output kickstart for kickstart installations won't always
have "firstboot --disable", which was not in the input kickstart
and overall makes more sense.

Also only try to enable the initial-setup.service if the unit file
exists to avoid bogus error messages.