Change PayloadRequirement to a DBus structure
Changing how the `strong` property is stored. It was stored in the named
tuple list with the reasons but it could be read only as one value. This
only complicates code. Store strong as single boolean value.

This strong value will be set to True if any of the reasons are strong
immediately when the reason is added.