Remove apply part of the payload Requirements
The new solution (modules) will not use signal and registration of the
callback. Instead it will just read the data when required by the
payload.

There is no need to have the callback when the only time it the callback
is registered is from a DNF payload and the only time the callback is
called is also from a DNF payload. In that case there is no reason to
complicate code and API with a signal.